suppresion table inutile, modification type date de booking en type datetime

+++ src/Entity/Dishes.php
@@ -35,10 +35,9 @@ class Dishes
 
     #[ORM\Column]
     #[Assert\PositiveOrZero(message: 'Le prix ne peut pas être vide ou inférieur à zéro')]
-    private ?int $price = null;
+    private ?float $price = null;
 
     #[ORM\Column]
-    #[Assert\NotBlank(message: 'Le plat préféré ne peut pas être vide')]
     private ?bool $bestDishes = null;
 
     /**
@@ -84,12 +83,12 @@ class Dishes
         return $this;
     }
 
-    public function getPrice(): ?int
+    public function getPrice(): ?float
     {
         return $this->price;
     }
 
-    public function setPrice(int $price): static
+    public function setPrice(float $price): static
     {
         $this->price = $price;
